refactor: split entity-detail-components.tsx into domain-focused modules
The 568-line grab-bag file was mixing verdict components, person cards,
funding rows, fact display, section headers, and collection tables.
Split into 7 focused files with a barrel re-export for backward compat.
